Move TemplateHelpers to Helpers.
