+++ lib/vedeu/support/all.rb
@@ -1,7 +1,6 @@
 require_relative 'common'
 require_relative 'options'
 require_relative 'template'
-require_relative 'template_helpers'
 require_relative 'terminal'
 require_relative 'timer'
 require_relative 'visibility'
