+++ lib/vedeu/templating/all.rb
@@ -1,2 +1,3 @@
+require_relative 'helpers'
 require_relative 'directive'
 require_relative 'preprocessor'
