+++ lib/vedeu/templating/directive.rb
@@ -5,6 +5,8 @@ module Vedeu
     # Converts a directive found in a template into a Vedeu::Stream object.
     class Directive
 
+      include Vedeu::Templating::Helpers
+
       # @param code [String]
       # @return [Vedeu::Stream]
       def self.process(code)
@@ -28,14 +30,6 @@ module Vedeu
       attr_reader :code
 
       private
-
-      # @param attributes [Hash]
-      # @param block [Proc]
-      # @return [Vedeu::Stream]
-      def colour(attributes = {}, &block)
-        Vedeu::Stream.new(colour: Vedeu::Colour.coerce(attributes),
-                          value:  block.call)
-      end
 
       # @return [Proc]
       def proc
